requirements.txt
boto3
yml
pyyaml
pandas
hurry.filesize
streamlit
Pillow